Automatically configure the web site to use the user's culture information that is passed as an HTTP header.

+++ OdeToFood/Controllers/HomeController.cs
@@ -26,7 +26,7 @@ namespace OdeToFood.Controllers
             return Json(model, JsonRequestBehavior.AllowGet);
         }
 
-        [OutputCache(CacheProfile="Long", VaryByHeader="X-Requested-With", Location = OutputCacheLocation.Server)]
+        [OutputCache(CacheProfile="Short", VaryByHeader="X-Requested-With", Location = OutputCacheLocation.Server)]
         public ActionResult Index(string searchTerm = null, int page = 1)
         {
             var model =
